requirements.txt
chardet==5.2.0
chroma-hnswlib==0.7.3
chromadb==0.5.2
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.1
huggingface-hub==0.23.4
langchain==0.2.5
langchain-chroma==0.1.1
langchain-community==0.2.5
langchain-core==0.2.11
langchain-google-genai==1.0.6
langchain-openai==0.1.8
langchain-text-splitters==0.2.2
langchainhub==0.1.20
langsmith==0.1.93
openai==1.35.3
python-dotenv==1.0.1
streamlit==1.37.1
